requirements.txt
absl-py==0.9.0
alabaster==0.7.12
anaconda-client==1.7.2
anaconda-navigator==1.9.7
anaconda-project==0.8.3
antlr4-python3-runtime==4.8
argh==0.26.2
asn1crypto==1.3.0
astroid==2.4.1
astropy==4.0.1.post1
astunparse==1.6.3
atomicwrites==1.4.0
attrs==19.3.0
autopep8==1.4.4
Babel==2.8.0
backcall==0.1.0
backports.functools-lru-cache==1.6.1
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
backports.tempfile==1.0
backports.weakref==1.0.post1
Beaker==1.11.0
beautifulsoup4==4.9.0
bitarray==1.2.1
bkcharts==0.2
bleach==3.1.4
bokeh==2.0.2
boto==2.49.0
Bottleneck==1.3.2
cachetools==4.1.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cloudpickle==1.4.1
clyent==1.2.2
colorama==0.4.3
conda==4.8.3
conda-build==3.18.8
conda-package-handling==1.7.0
conda-verify==3.4.2
configparser==5.0.1
contextlib2==0.6.0.post1
crcmod==1.7
cryptography==2.9.2
cycler==0.10.0
Cython==0.29.17
cytoolz==0.10.1
dask==2.17.2
dataclasses==0.6
datasets==1.2.1
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20181111
dill==0.3.2
distributed==2.17.0
docker-pycreds==0.4.0
docutils==0.16
entrypoints==0.3
et-xmlfile==1.0.1
fairseq==0.10.2
fastcache==1.1.0
filelock==3.0.12
flake8==3.8.2
Flask==1.1.2
Flask-SocketIO==4.2.1
fsspec==0.7.4
future==0.18.2
gast==0.3.3
gevent==1.4.0
gitdb==4.0.5
GitPython==3.1.12
glob2==0.7
gmpy2==2.0.8
google-auth==1.20.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
googleapis-common-protos==1.52.0
greenlet==0.4.15
grpcio==1.31.0
h5py==2.10.0
HeapDict==1.0.1
html5lib==1.0.1
hydra-core==1.0.6
hypothesis==5.11.0
idna==2.9
imageio==2.8.0
imagesize==1.2.0
importlib-metadata==1.6.0
importlib-resources==5.1.0
intervaltree==3.0.2
ipdb==0.13.4
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.15.2
jeepney==0.4.3
Jinja2==2.11.2
joblib==0.15.1
json5==0.9.4
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
jupyterlab==1.2.6
jupyterlab-server==1.0.9
Keras-Preprocessing==1.1.2
keyring==21.1.1
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
libarchive-c==2.9
llvmlite==0.29.0
locket==0.2.0
lxml==4.5.0
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.1.3
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.0.15
mkl-random==1.1.1
mkl-service==2.3.0
mock==4.0.2
more-itertools==8.3.0
mpmath==1.1.0
msgpack==1.0.0
multipledispatch==0.6.0
multiprocess==0.70.11.1
navigator-updater==0.2.1
nbconvert==5.6.1
nbformat==5.0.6
networkx==2.4
nltk==3.4.5
nose==1.3.7
notebook==6.0.3
numba==0.44.1
numexpr==2.7.1
numpy==1.18.1
numpydoc==0.9.2
oauthlib==3.1.0
olefile==0.46
omegaconf==2.0.6
openpyxl==3.0.3
opt-einsum==3.3.0
packaging==20.3
pandas==1.0.3
pandocfilters==1.4.2
parso==0.5.2
partd==1.1.0
path==13.2.0
path.py==12.0.2
pathlib2==2.3.5
pathtools==0.1.2
patsy==0.5.1
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
pkginfo==1.5.0.1
pluggy==0.13.1
ply==3.11
portalocker==2.2.1
prometheus-client==0.7.1
promise==2.3
prompt-toolkit==3.0.4
protobuf==3.12.4
psutil==5.7.0
ptyprocess==0.6.0
py==1.8.1
pyarrow==3.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycosat==0.6.3
pycparser==2.20
pycrypto==2.6.1
pycurl==7.43.0.5
pydocstyle==4.0.1
pyflakes==2.2.0
Pygments==2.6.1
pylint==2.5.2
pyodbc===4.0.0-unsupported
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrsistent==0.16.0
PySocks==1.7.1
pytest==5.4.2
pytest-arraydiff==0.3
pytest-astropy==0.8.0
pytest-astropy-header==0.1.2
pytest-doctestplus==0.5.0
pytest-openfiles==0.5.0
pytest-remotedata==0.3.2
python-dateutil==2.8.1
python-engineio==3.12.1
python-jsonrpc-server==0.3.4
python-language-server==0.31.9
python-socketio==4.5.0
pytz==2020.1
PyWavelets==1.1.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==18.1.1
QDarkStyle==2.8.1
QtAwesome==0.7.0
qtconsole==4.7.4
QtPy==1.9.0
regex==2020.7.14
requests==2.23.0
requests-oauthlib==1.3.0
rope==0.17.0
rouge-score==0.0.4
rsa==4.6
Rtree==0.9.4
ruamel-yaml==0.15.87
sacrebleu==1.5.0
sacremoses==0.0.43
scikit-image==0.16.2
scikit-learn==0.22.1
scipy==1.4.1
seaborn==0.10.1
SecretStorage==3.1.2
Send2Trash==1.5.0
sentencepiece==0.1.91
sentry-sdk==0.19.5
shortuuid==1.0.1
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.14.0
smmap==3.0.5
snowballstemmer==2.0.0
sortedcollections==1.1.2
sortedcontainers==2.1.0
soupsieve==2.0.1
Sphinx==3.0.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sphinxcontrib-websupport==1.2.1
spyder==4.1.3
spyder-kernels==1.9.1
SQLAlchemy==1.3.17
statsmodels==0.11.1
subprocess32==3.5.4
sympy==1.5.1
tables==3.6.1
tblib==1.6.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow-datasets==3.2.1
tensorflow-estimator==2.3.0
tensorflow-gpu==2.3.0
tensorflow-hub==0.8.0
tensorflow-metadata==0.22.2
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tf-sentencepiece==0.1.90
tokenizers==0.9.4
toml==0.10.0
toolz==0.10.0
torch==1.7.0
tornado==6.0.4
tqdm==4.46.0
traitlets==4.3.3
transformers==4.2.2
typed-ast==1.4.1
typing-extensions==3.7.4.1
ujson==1.35
unicodecsv==0.14.1
urllib3==1.25.8
wandb==0.10.17
watchdog==0.10.2
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wikimapper==0.1.5
wrapt==1.11.2
wurlitzer==2.0.0
xlrd==1.2.0
XlsxWriter==1.2.8
xlwt==1.3.0
xxhash==2.0.0
yapf==0.28.0
zict==2.0.0
zipp==3.1.0
